refactor : api 함수 분리
